requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
asgiref==3.5.0
    # via django
asttokens==2.0.5
    # via stack-data
backcall==0.2.0
    # via ipython
brotli==1.0.9
    # via geventhttpclient
certifi==2021.10.8
    # via
    #   geventhttpclient
    #   requests
charset-normalizer==2.0.12
    # via requests
click==8.1.3
    # via flask
configargparse==1.5.3
    # via locust
decorator==5.1.1
    # via ipython
django==3.2.13
    # via
    #   -r requirements.in
    #   django-debug-toolbar
    #   djangorestframework
django-debug-toolbar==3.2.4
    # via -r requirements.in
djangorestframework==3.13.1
    # via -r requirements.in
executing==0.8.3
    # via stack-data
flask==2.1.2
    # via
    #   flask-basicauth
    #   flask-cors
    #   locust
flask-basicauth==0.2.0
    # via locust
flask-cors==3.0.10
    # via locust
gevent==21.12.0
    # via
    #   geventhttpclient
    #   locust
geventhttpclient==1.5.3
    # via locust
greenlet==1.1.2
    # via gevent
gunicorn==20.1.0
    # via -r requirements.in
idna==3.3
    # via requests
importlib-metadata==4.11.3
    # via flask
ipython==8.2.0
    # via -r requirements.in
itsdangerous==2.1.2
    # via flask
jedi==0.18.1
    # via ipython
jinja2==3.0.3
    # via
    #   flask
    #   locust
locust==2.8.6
    # via -r requirements.in
markupsafe==2.1.1
    # via jinja2
matplotlib-inline==0.1.3
    # via ipython
msgpack==1.0.3
    # via locust
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
prompt-toolkit==3.0.29
    # via ipython
psutil==5.9.0
    # via locust
psycopg2==2.9.3
    # via -r requirements.in
psycopg2-binary==2.9.3
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pygments==2.12.0
    # via ipython
python-dotenv==0.20.0
    # via -r requirements.in
pytz==2022.1
    # via
    #   django
    #   djangorestframework
pyzmq==22.3.0
    # via locust
requests==2.27.1
    # via
    #   -r requirements.in
    #   locust
roundrobin==0.0.2
    # via locust
six==1.16.0
    # via
    #   asttokens
    #   flask-cors
    #   geventhttpclient
sqlparse==0.4.2
    # via
    #   django
    #   django-debug-toolbar
stack-data==0.2.0
    # via ipython
traitlets==5.1.1
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.2.0
    # via locust
urllib3==1.26.9
    # via requests
wcwidth==0.2.5
    # via prompt-toolkit
werkzeug==2.1.2
    # via
    #   flask
    #   locust
zipp==3.8.0
    # via importlib-metadata
zope-event==4.5.0
    # via gevent
zope-interface==5.4.0
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
